+++ R/tabyl.R
@@ -227,6 +227,11 @@ tabyl_2way <- function(dat, var1, var2, show_na = TRUE, show_missing_levels = TR
 # a list of two-way frequency tables, split into a list on a third variable
 tabyl_3way <- function(dat, var1, var2, var3, show_na = TRUE, show_missing_levels = TRUE) {
   dat <- dplyr::select(dat, !! var1, !! var2, !! var3)
+  
+  # Keep factor levels for ordering the list at the end
+  if(is.factor(dat[[3]])){
+    third_levels <- levels(dat[[3]])
+  }
   dat[[3]] <- as.character(dat[[3]]) # don't want empty factor levels in the result list - they would be empty data.frames
 
   # grab class of 1st variable to restore it later
@@ -258,6 +263,13 @@ tabyl_3way <- function(dat, var1, var2, var3, show_na = TRUE, show_missing_level
     purrr::map(tabyl_2way, var1, var2, show_na = show_na, show_missing_levels = show_missing_levels) %>%
     purrr::map(reset_1st_col_status, col1_class, col1_levels) # reset class of var in 1st col to its input class, #168
 
+  # reorder when var 3 is a factor, per #250
+  if(exists("third_levels")){
+    result <- result[order(third_levels[third_levels %in% unique(dat[[3]])])] 
+  # missing levels will have been dropped at conversion to factor, good
+  #CHECK BEHAVIOR WITH NAs - tough to do until #274 is properly fixed.  NA level should appear if show_na = TRUE
+  }
+  
   result
 }